+++ tensquare_recruit/src/main/java/com/tensquare/recruit/entity/TbEnterprise.java
@@ -19,14 +19,50 @@ public class TbEnterprise implements Serializable {
 
   @Id
   private String id;
+
+  /**
+   * 企业名称
+   */
   private String name;
+
+  /**
+   * 企业简介
+   */
   private String summary;
+
+  /**
+   * 企业地址
+   */
   private String address;
+
+  /**
+   * 标签列表（用逗号分隔）
+   */
   private String labels;
+
+  /**
+   * 企业位置坐标（经度，纬度）
+   */
   private String coordinate;
+
+  /**
+   * 是否热门（0:非热门 1:热门）
+   */
   private String ishot;
+
+  /**
+   * LOGO
+   */
   private String logo;
-  private long jobcount;
+
+  /**
+   * 职位数
+   */
+  private Integer jobcount;
+
+  /**
+   * URL
+   */
   private String url;
 
   @Override
